test: move shared token and character helpers into Pest.php
Relocate trackedToken() and characterFor() from ConnectorTest.php to Pest.php
alongside fakeCharacter() so every test file can use them. This lets
CursorPaginationTest run standalone instead of depending on ConnectorTest
being loaded first.
